add nolinux stubs
